+++ configs/regression/_base_/models/resnet/resnet18_cifar.py
@@ -9,7 +9,7 @@ model = dict(
         out_indices=(3,),  # no conv-1, x-1: stage-x
         style='pytorch'),
     head=dict(
-        type='ClsHead',
-        loss=dict(type='CrossEntropyLoss', loss_weight=1.0),
-        with_avg_pool=True, in_channels=512, num_classes=1000)
+        type='RegHead',
+        loss=dict(type='RegressionLoss', mode="l1_loss", loss_weight=1.0),
+        with_avg_pool=True, in_channels=512, out_channels=1)
 )