+++ configs/regression/_base_/models/vision_transformer/vit_small_p16_sz224.py
@@ -1,10 +1,7 @@
 # model settings
 model = dict(
-    type='MixUpClassification',
+    type='Classification',
     pretrained=None,
-    alpha=[0.8, 1.0,],
-    mix_mode=["mixup", "cutmix",],
-    mix_args=dict(),
     backbone=dict(
         type='VisionTransformer',
         arch='base',
@@ -12,8 +9,7 @@ model = dict(
         drop_path_rate=0.1,
     ),
     head=dict(
-        type='VisionTransformerClsHead',
-        loss=dict(type='LabelSmoothLoss',
-            label_smooth_val=0.1, num_classes=1000, mode='original', loss_weight=1.0),
-        in_channels=768, num_classes=1000)
+        type='RegHead',
+        loss=dict(type='RegressionLoss', mode="l1_loss", loss_weight=1.0),
+        with_avg_pool=False, in_channels=768, out_channels=1)
 )