+++ configs/regression/agedb/basic/convnext_t_l1_sz224_adamw_4xbs64_ep400.py
@@ -1,0 +1,38 @@
+_base_ = [
+    '../../_base_/models/convnext/convnext_tiny.py',
+    '../../_base_/datasets/agedb/basic_sz224_4xbs64.py',
+    '../../_base_/default_runtime.py',
+]
+
+# additional hooks
+update_interval = 1
+
+# optimizer
+optimizer = dict(
+    type='AdamW',
+    lr=1e-3,
+    weight_decay=0.05, eps=1e-8, betas=(0.9, 0.999),
+    paramwise_options={
+        '(bn|ln|gn)(\d+)?.(weight|bias)': dict(weight_decay=0.),
+        'norm': dict(weight_decay=0.),
+        'bias': dict(weight_decay=0.),
+        'gamma': dict(weight_decay=0.),
+    })
+
+# fp16
+use_fp16 = False
+fp16 = dict(type='mmcv', loss_scale='dynamic')
+optimizer_config = dict(
+    grad_clip=dict(max_norm=5.0), update_interval=update_interval)
+
+# learning policy
+lr_config = dict(
+    policy='CosineAnnealing',
+    by_epoch=False, min_lr=1e-6,
+    warmup='linear',
+    warmup_iters=5, warmup_by_epoch=True,  # warmup 5 epochs.
+    warmup_ratio=1e-6,
+)
+
+# runtime settings
+runner = dict(type='EpochBasedRunner', max_epochs=400)