+++ configs/regression/imdb_wiki/basic/r50_l1_sz224_adamw_4xbs64_ep200.py
@@ -1,0 +1,20 @@
+_base_ = [
+    '../../_base_/models/resnet/resnet50.py',
+    '../../_base_/datasets/imdb_wiki/basic_sz224_4xbs64.py',
+    '../../_base_/default_runtime.py',
+]
+
+# additional hooks
+update_interval = 1
+
+# optimizer
+optimizer = dict(
+    type='AdamW', lr=1e-3,
+    weight_decay=0.05, eps=1e-8, betas=(0.9, 0.999))
+optimizer_config = dict(grad_clip=None)
+
+# learning policy
+lr_config = dict(policy='CosineAnnealing', min_lr=1e-6)
+
+# runtime settings
+runner = dict(type='EpochBasedRunner', max_epochs=200)