+++ docs/en/changelog.md
@@ -8,7 +8,7 @@ Bump version to V0.2.8 with new features in [MMPreTrain](https://github.com/open
 
 - Support more backbone architectures, including [MobileNetV3](https://arxiv.org/abs/1905.02244), [EfficientNetV2](https://arxiv.org/abs/2104.00298), [HRNet](https://arxiv.org/abs/1908.07919), [CSPNet](https://arxiv.org/abs/1911.11929), [LeViT](https://arxiv.org/abs/2104.01136), [MobileViT](http://arxiv.org/abs/2110.02178), [DaViT](https://arxiv.org/abs/2204.03645), and [MobileOne](http://arxiv.org/abs/2206.04040), etc.
 - Support CIFAR-100 benchmarks of Metaformer architectures and Mixup variants with Transformers, detailed in [cifar100/advanced](https://github.com/Westlake-AI/openmixup/blob/main/configs/classification/cifar100/advanced) and [cifar100/mixups](https://github.com/Westlake-AI/openmixup/blob/main/configs/classification/cifar100/mixups). Models and logs of various CIFAR-100 mixup benchmarks are on updating.
-- Support regression tasks with relavent datasets, metrics, and [configs](https://github.com/Westlake-AI/openmixup/blob/main/configs/regression).
+- Support regression tasks with relavent datasets, metrics, and [configs](https://github.com/Westlake-AI/openmixup/blob/main/configs/regression). Datasets include [AgeDB](https://ieeexplore.ieee.org/document/8014984), [IMDB-WIKI](https://link.springer.com/article/10.1007/s11263-016-0940-3), and [RCFMNIST](https://arxiv.org/abs/2210.05775).
 
 ### Update Documents
 
@@ -17,6 +17,7 @@ Bump version to V0.2.8 with new features in [MMPreTrain](https://github.com/open
 ### Bug Fixes
 
 - Fix the `by_epoch` setting in `CustomSchedulerHook` and update `DecoupleMix` in `soft_mix_cross_entropy` to support label smoothing settings.
+- Fix bugs of Vision Transformers in [cls_mixup_head](https://github.com/Westlake-AI/openmixup/blob/main/openmixup/models/heads/cls_mixup_head.py) and [reg_head](https://github.com/Westlake-AI/openmixup/blob/main/openmixup/models/heads/reg_head.py).
 
 ### v0.2.7 (16/12/2022)